GCC/Linux build fixes
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6107 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/unix/unix.cpp
@@ -26,7 +26,7 @@ void sys_display_msgw(const wchar_t* caption, const wchar_t* msg)
 	fwprintf(stderr, L"%ls: %ls\n", caption, msg);
 }
 
-ErrorReaction sys_display_error(const wchar_t* text, size_t flags)
+ErrorReaction sys_display_error(const wchar_t* text, int flags)
 {
 	printf("%ls\n\n", text);
